+++ server/queries/trips.sql
@@ -1,8 +1,8 @@
 -- name: InsertTrip :one
 INSERT INTO
-    trips (name, description)
+    trips (name, description, owner_id)
 VALUES
-    ($1, $2) RETURNING *;
+    (@name, @description, @owner_id) RETURNING *;
 
 -- name: GetTripById :one
 SELECT
@@ -10,16 +10,51 @@ SELECT
 FROM
     trips
 WHERE
-    id = @id;
+    id = @id
+    AND (
+        owner_id = @user_id
+        OR id IN (
+            SELECT
+                trip_id
+            FROM
+                trip_members
+            WHERE
+                user_id = @user_id
+        )
+    );
 
 -- name: DeleteTripById :exec
 DELETE FROM
     trips
 WHERE
-    id = @id;
+    id = @id
+    AND owner_id = @user_id;
 
 -- name: ListTrips :many
 SELECT
     *
 FROM
-    trips;
+    trips
+WHERE
+    owner_id = @user_id
+    OR id IN (
+        SELECT
+            trip_id
+        FROM
+            trip_members
+        WHERE
+            user_id = @user_id
+    );
+
+-- name: AddTripMember :exec
+INSERT INTO
+    trip_members (trip_id, user_id)
+VALUES
+    (@trip_id, @user_id);
+
+-- name: RemoveTripMember :exec
+DELETE FROM
+    trip_members
+WHERE
+    trip_id = @trip_id
+    AND user_id = @user_id;
